Requisicao com os ciclos e suas areas

+++ Censo.API/Model/dados/Enquadramento.cs
@@ -2,7 +2,7 @@ namespace Censo.API.Model.dados
 {
     public class Enquadramento
     {
-        public long CodEnq { get; set; }
+        public double CodEnq { get; set; }
         public string NomEnq { get; set; }
         
     }
